requirements.txt
annotated-types==0.7.0
attrs==20.3.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
check50==3.3.11
cryptography==43.0.1
idna==3.10
jellyfish==0.11.2
Jinja2==3.1.4
lib50==3.0.11
MarkupSafe==2.1.5
packaging==24.1
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
soupsieve==2.6
submit50==3.1.5
termcolor==1.1.0
typing_extensions==4.12.2
urllib3==2.2.3
